chore: make the template docs view the default
