+++ site/src/pages/TemplatePage/TemplateStats.stories.tsx
