+++ site/src/pages/TemplatePage/TemplateStats.tsx
